pyproject.toml
[project]
name = "lolmcp"
version = "0.1.0"
description = "Autonomous iOS security research platform fusing Frida, Objection, and MITMProxy MCP."
readme = "README.md"
requires-python = ">=3.12,<3.14"
license = { text = "MIT" }
authors = [{ name = "xtofuub" }]

dependencies = [
    "pydantic>=2.6,<3",
    "structlog>=24.1,<26",
    "httpx>=0.27,<1",
    "anyio>=4.3,<5",
    "click>=8.1,<9",
    "PyYAML>=6.0,<7",
    "Jinja2>=3.1,<4",
    "aiofiles>=23.2,<25",
    "python-ulid>=2.2,<4",
    "PyJWT>=2.8,<3",
    "frida>=16.5,<18",
    "frida-tools>=13.0,<15",
    "objection>=1.11.0,<2",
    "mitmproxy>=12.0.0,<13",
    "mcp>=1.2,<2",
    "rich>=13.7,<15",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9",
    "pytest-asyncio>=0.23,<1",
    "pytest-cov>=5.0,<6",
    "ruff>=0.4,<1",
    "mypy>=1.10,<2",
    "types-PyYAML>=6.0,<7",
]

[project.scripts]
lolmcp = "agent.cli:main"

[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent", "api", "frida_layer", "objection_layer", "mitm"]

[tool.ruff]
line-length = 100
target-version = "py312"
extend-exclude = ["mitm/vendor", "skills/_upstream"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "ASYNC", "S"]
ignore = ["S101"]

[tool.mypy]
python_version = "3.12"
strict = true
exclude = ["mitm/vendor/", "skills/_upstream/"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
